requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
distro==1.9.0
Django==4.2.28
exceptiongroup==1.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.13.0
msgpack==1.1.2
openai==2.21.0
packaging==26.0
pillow==11.3.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dotenv==1.2.1
razorpay==2.0.0
redis==7.0.1
requests==2.32.5
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
tomli==2.4.0
tqdm==4.67.3
Twisted==25.5.0
txaio==23.6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
zope.interface==8.0.1
